+++ _source/alias/media.sh
@@ -5,8 +5,7 @@ if [[ -f /usr/bin/cyanrip  ]];
 then
   alias ripCdAsMp3=net_bazzline_media_rip_cd_as_mp3
 fi
-
-if [[ -f /usr/bin/vobcopy ]];
+if [[ -f /usr/bin/vobcopy ]] && [[ -f /usr/bin/ffmpeg ]];
 then
-  alias ripDvdAsVob='vobcopy -M'
+  alias ripDvdAs265Mkv=net_bazzline_media_rip_dvd_to_x265
 fi